+++ packages/press-ui/src/packages/common/utils/rpx.d.ts
@@ -1,2 +1,1 @@
 export const rpx2px: (rpx: number) => number;
-export const systemInfo: WechatMiniprogram.SystemInfo;