+++ packages/press-ui/src/packages/common/utils/rpx.js
@@ -1,3 +1,5 @@
-export const systemInfo = wx.getSystemInfoSync();
+import { getWindowWidth } from '../dom/rect';
+
+const systemInfo = getWindowWidth();
 
 export const rpx2px = rpx => Math.floor((systemInfo.windowWidth * rpx) / 750);