+++ packages/press-ui/src/packages/press-index-bar/demo.vue
@@ -5,7 +5,6 @@
   >
     <press-index-bar
       :sticky-offset-top="stickyOffsetTop"
-      :window-top="44"
       @select="onSelect"
     >
       <div